Impl. Workaround for Clearing TLST ID After MyMusic
In some situations the ID used for determining whether to reload the TLST gets cleared after exiting MyMusic, which in our use-case broke support for tracks that relied on their TLST to load correctly. This provides a mechanism to forcibly *restore* that ID if and only if the cleared ID was the "Menu ID" (including our spoofed one for keeping the loaded TLST).