Impl. Fallback for Failed Song Start
Songs currently begin playing when you leave a Stage's My Music page, which triggers the code to make it the active menu track. This is a problem though, in that doing this, then moving to another stage and backing out will attempt to initialize the same song, which will cause nothing to play if the newly loaded TLST doesn't have the original song's ID. This remedies this by just ensuring that, if we back out of a My Music page with nothing playing, we always return to Menu Music!